innit Normal

+++ math/probability/binomial.py
@@ -41,7 +41,7 @@ class Binomial:
                 raise ValueError("data must contain multiple values")
 
             p = sum(data) / len(data)
-            n = round(sum(data) / p)
+            n = int(round(sum(data) / p))
             p = sum(data) / n
-            self.n = int(n)
-            self.p = float(p)
+            self.n = n
+            self.p = p
